test: move selecting items tests to own file
